taskmanager/Pie/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
flake8==6.1.0
isort==5.12.0
mccabe==0.7.0
packaging==23.1
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
gunicorn==20.1.0
openpyxl